Restructure layout of app
