+++ frontend/src/App.css
@@ -1,6 +1,31 @@
-#root {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 2rem;
-  text-align: center;
+/* Layout */
+
+.page {
+  min-height: 100vh;
+  width: 100%;
+  position: relative;
+}
+
+/* Header */
+.header-container {
+  background-color: #14203CFA;
+  color: white;
+  min-height: 8vh;
+  padding: 1rem 2rem;
+  gap: 1.5rem;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.page-contents {
+  box-sizing: border-box;
+  min-height: 75vh;
+  width: 100%;
+  padding: 4rem;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 1.5rem;
 }
