+++ frontend/src/layout/Layout.jsx
@@ -1,0 +1,13 @@
+import { Outlet } from 'react-router-dom';
+import Header from './Header';
+
+const Layout = () => {
+  return (
+    <div className="page">
+      <Header />
+      <Outlet />
+    </div>
+  );
+};
+
+export default Layout;
